+++ src/NSchema/Configuration/State/S3StateConfig.cs
@@ -1,5 +1,4 @@
 using NSchema.Configuration.Ddl;
-using NSchema.Schema.Ddl;
 
 namespace NSchema.Configuration.State;